VFB-117: Add subscription to packing slot table to update rows
