+++ src/app/admin/page.tsx
@@ -8,6 +8,7 @@ import { User } from "@supabase/gotrue-js";
 import { DatabaseError } from "@/app/errorClasses";
 import { Schema } from "@/databaseUtils";
 import { PackingSlotRow } from "@/app/admin/packingSlotsTable/PackingSlotsTable";
+import { fetchPackingSlots } from "@/app/admin/packingSlotsTable/PackingSlotActions";
 
 // disables caching
 export const revalidate = 0;
@@ -53,31 +54,11 @@ const getCollectionCentres = async (): Promise<Schema["collection_centres"][]> =
     return data;
 };
 
-const getPackingSlots = async (): Promise<PackingSlotRow[]> => {
-    const supabase = getSupabaseServerComponentClient();
-    const { data, error } = await supabase.from("packing_slots").select().order("order");
-    if (error) {
-        throw new DatabaseError("fetch", "packing slots");
-    }
-
-    const processedData = data.map((row): PackingSlotRow => {
-        return {
-            name: row.name,
-            id: row.primary_key,
-            is_hidden: row.is_hidden,
-            order: row.order,
-            isNew: false,
-        };
-    });
-
-    return processedData;
-};
-
 const Admin = async (): Promise<ReactElement> => {
     const [userData, collectionCentreData, packingSlotData] = await Promise.all([
         getUsers(),
         getCollectionCentres(),
-        getPackingSlots(),
+        fetchPackingSlots(),
     ]);
 
     return (
